rename collecitve.py, test=develop
